+++ fisk-factory-model/fisk-factory-model-web/src/main/java/com/fisk/datamodel/service/impl/fact/FactImpl.java
@@ -43,13 +43,13 @@ import com.fisk.datamodel.service.impl.BusinessAreaImpl;
 import com.fisk.datamodel.service.impl.dimension.DimensionImpl;
 import com.fisk.datamodel.vo.DataModelTableVO;
 import com.fisk.datamodel.vo.DataModelVO;
-import com.fisk.system.client.UserClient;
 import com.fisk.task.client.PublishTaskClient;
 import com.fisk.task.dto.pgsql.PgsqlDelTableDTO;
 import com.fisk.task.dto.pgsql.TableListDTO;
 import com.fisk.task.enums.DataClassifyEnum;
 import com.fisk.task.enums.OlapTableEnum;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
@@ -84,11 +84,12 @@ public class FactImpl extends ServiceImpl<FactMapper, FactPO> implements IFact {
     @Resource
     UserHelper userHelper;
     @Resource
-    UserClient userClient;
-    @Resource
     DimensionImpl dimensionImpl;
     @Resource
     DataManageClient dataManageClient;
+
+    @Value("${spring.open-metadata}")
+    private Boolean openMetadata;
 
     @Override
     public ResultEnum addFact(FactDTO dto) {
@@ -162,9 +163,12 @@ public class FactImpl extends ServiceImpl<FactMapper, FactPO> implements IFact {
                 if (dataSourceConfigOlap != null && !CollectionUtils.isEmpty(dataSourceConfigOlap.dbList)) {
                     delQualifiedName.add(dataSourceConfigOlap.dbList.get(0).qualifiedName + "_" + DataModelTableTypeEnum.DORIS_FACT.getValue() + "_" + id);
                 }
-                deleteDto.qualifiedNames = delQualifiedName;
-                deleteDto.classifications = businessArea.getBusinessName();
-                dataManageClient.deleteMetaData(deleteDto);
+
+                if (openMetadata) {
+                    deleteDto.qualifiedNames = delQualifiedName;
+                    deleteDto.classifications = businessArea.getBusinessName();
+                    dataManageClient.deleteMetaData(deleteDto);
+                }
             }
 
             return flat > 0 ? ResultEnum.SUCCESS : ResultEnum.SAVE_DATA_ERROR;
@@ -366,20 +370,22 @@ public class FactImpl extends ServiceImpl<FactMapper, FactPO> implements IFact {
         data.dbList.get(0).tableList = tableList;
         list.add(data);
 
-        //修改元数据
-        ExecutorService cachedThreadPool = Executors.newCachedThreadPool();
-        cachedThreadPool.execute(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    // 更新元数据内容
-                    log.info("维度表构建元数据实时同步数据对象开始.........: 参数为: {}", JSON.toJSONString(list));
-                    dataManageClient.consumeMetaData(list);
-                } catch (Exception e) {
-                    log.error("【dataManageClient.MetaData()】方法报错,ex", e);
+        if (openMetadata) {
+            //修改元数据
+            ExecutorService cachedThreadPool = Executors.newCachedThreadPool();
+            cachedThreadPool.execute(new Runnable() {
+                @Override
+                public void run() {
+                    try {
+                        // 更新元数据内容
+                        log.info("维度表构建元数据实时同步数据对象开始.........: 参数为: {}", JSON.toJSONString(list));
+                        dataManageClient.consumeMetaData(list);
+                    } catch (Exception e) {
+                        log.error("【dataManageClient.MetaData()】方法报错,ex", e);
+                    }
                 }
-            }
-        });
+            });
+        }
 
     }
